Fix remaining PR #98 review feedback
Additional fixes for unresolved code review comments:

1. **milliseconds format specifier** (gateway.cpp:728, 743)
   - Changed from 'long' with '%ld' to 'long long' with '%lld'
   - Fixes undefined behavior: milliseconds::count() returns long long
   - Removes unnecessary narrowing cast

2. **resolveLogFilePathFromEnvironment() return type** (logger.cpp:95-103)
   - Changed from std::optional<std::string> to std::string
   - Returns empty string when env var not set (cleaner API)
   - Simplifies caller logic: empty() check instead of has_value()
   - Updated call site and pointer dereference (->c_str() to .c_str())

All 129 unit tests passing. Clang-format compliant.

+++ src/gateway.cpp
@@ -724,12 +724,12 @@ public:
         auto t0_unsub = std::chrono::steady_clock::now();
         auto future_unsub = request(event, params);
         constexpr auto kUnsubscribeAckTimeout = std::chrono::milliseconds(50);
-        long unsub_ms = 0;
+        long long unsub_ms = 0;
         if (future_unsub.wait_for(kUnsubscribeAckTimeout) == std::future_status::ready)
         {
-            unsub_ms = static_cast<long>(
-                std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::steady_clock::now() - t0_unsub).count());
-            FIREBOLT_LOG_DEBUG("Gateway", "[unsubscribe] ACK received after %ld ms", unsub_ms);
+            unsub_ms =
+                std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::steady_clock::now() - t0_unsub).count();
+            FIREBOLT_LOG_DEBUG("Gateway", "[unsubscribe] ACK received after %lld ms", unsub_ms);
             auto result = future_unsub.get();
             if (!result)
             {
@@ -738,9 +738,9 @@ public:
         }
         else
         {
-            unsub_ms = static_cast<long>(
-                std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::steady_clock::now() - t0_unsub).count());
-            FIREBOLT_LOG_DEBUG("Gateway", "[unsubscribe] ACK timed out after %ld ms (server unresponsive)", unsub_ms);
+            unsub_ms =
+                std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::steady_clock::now() - t0_unsub).count();
+            FIREBOLT_LOG_DEBUG("Gateway", "[unsubscribe] ACK timed out after %lld ms (server unresponsive)", unsub_ms);
         }
 
         return status;
